Module rand_split::chaskeyrng

A splittable pseudorandom generator based on the Chaskey MAC. This is not intended to be a cryptographically secure PRNG.

Like SipRng, this is broadly modeled after Claessen and Pałka's splittable PRNGs, but with a different choice of cryptographic primitive.

References

Mouha, Nicky, Bart Mennik, Anthony Van Herrewege, Dai Watanabe, Bart Preneet and Ingrid Verbauwhede. 2014. "Chaskey: An Efficient MAC Algorithm for 32-bit Microcontrollers." Cryptology ePrint Archive, Report 2014/386.

Claessen, Koen and Michał H. Pałka. 2013. "Splittable Pseudorandom Number Generators using Cryptographic Hashing." Haskell '13: Proceedings of the 2013 ACM SIGPLAN symposium on Haskell, pp. 47-58.

Structs

ChaskeyPrf
A PRF taken off a ChaskeyRng.

ChaskeyRng
A splittable pseudorandom generator based on Chaskey.