Module rand_split::siprng

A splittable pseudo-random number generator based on the SipHash function. This is not intended to be a cryptographically secure PRNG.

This generator is broadly modeled after Claessen and Pałka's, a version of which is implemented in the Haskell tf-random library. Instead of the Skein hash function, however, we use SipHash as the pseudo-random function.

An outline of how this generator works, conceptually:

We SipHash as a pseudo-random function (PRF), keyed with the RNG seed (128 bits).

The generator (conceptually) records the history of operations that have been invoked on it and its "parents."

To generate random output, this history is interpreted as a string and hashed.

By computing the PRF incrementally this can be done in constant time and space. I.e., the generator's state is the intermediate state of hashing the prefix of the operation string that it has seen so far.

References

Aumasson, Jean-Philippe and Daniel J. Bernstein. 2012. "SipHash: a fast short-input PRF." Cryptology ePrint Archive, Report 2012/351.

Claessen, Koen and Michał H. Pałka. 2013. "Splittable Pseudorandom Number Generators using Cryptographic Hashing." Haskell '13: Proceedings of the 2013 ACM SIGPLAN symposium on Haskell, pp. 47-58.

Structs

SipPrf
A PRF taken off a SipRng.

SipRng
A splittable pseudorandom generator based on SipHash.