Trait rand_split::SplitPrf

pub trait SplitPrf<Rng> {
    fn call(&self, i: u64) -> Rng;
}

Pseudo-random functions ("PRFs") generated off a SplitRng.

A PRF is a factory of SplitRngs, whose initial states are determined by two values:

The state of the SplitRng when the PRF was created;

The argument to the PRF.

Note that while the term pseudo-random function has a technical meaning in cryptograpy, no security claim is implied here.

Required Methods

fn call(&self, i: u64) -> Rng