Trait rand_split::SplitRng

pub trait SplitRng: Rng + Sized {
    type Prf: SplitPrf<Self>;
    fn splitn(&mut self) -> Self::Prf;
    fn split(&mut self) -> Self;

    fn split_gen<A: SplitRand>(&mut self) -> A { ... }
}

A trait for splittable pseudo random generators.

Associated Types

type Prf: SplitPrf<Self>

The type of pseudo-random functions ("PRFs") produced off a SplitRng instance. A PRF is a factory of SplitRngs, whose initial states are determined by two values:

The state of the SplitRng when the PRF was created;

The argument to the PRF.

Note that while the term pseudo-random function has a technical meaning in cryptograpy, no security claim is implied here.

Required Methods

fn splitn(&mut self) -> Self::Prf

Split a pseudo-random function off this generator.

fn split(&mut self) -> Self

Split a second RNG off this one.

Provided Methods

fn split_gen<A: SplitRand>(&mut self) -> A