Crate rand_split

A crate to support splittable random number generators.

These are genertors that support a "split" operation that produces two or more "child" generators whose initial state depends on the source RNG's state, but produce outputs that are uncorrelated to each other. This can be useful in some contexts:

Functional programs sometimes make good use of splittable RNGs. The Haskell QuickCheck library is a notable example that uses a splittable, pure-functional PRNG to support random generation of deterministic functions.

Parallel programs may benefit from a splittable PRNG. The advantage isn't necessarily performance but rather reproducibility; it makes it easier to guarantee that using the same seed yields the same results even if the execution order is nondeterministic.

References:

Claessen, Koen and Michał H. Pałka. 2013. "Splittable Pseudorandom Number Generators using Cryptographic Hashing." Haskell '13: Proceedings of the 2013 ACM SIGPLAN symposium on Haskell, pp. 47-58.

Schaathun, Hans Georg. 2015. "Evaluation of Splittable Pseudo-Random Generators." Journal of Functional Programming, Vol. 25.

The Haskell tf-random library.

Modules

generic
A construction that turns a pair of a splittable and a sequential PRNG into a splittable PRNG. The intent of this is to gain the splittability of the former but retain the sequential generation speed of the latter.

siprng
A splittable pseudo-random number generator based on the SipHash function. This is not intended to be a cryptographically secure PRNG.

Macros

split_rand_seq_impl!
A macro that implements SplitRand sequentially for any type that has a Rand implementation, simply by using that. This is meant to be used for "atomic" types whose generation doesn't benefit from splittability.

Traits

SplitPrf
Pseudo-random functions ("PRFs") generated off a SplitRng.

SplitRand
A type that can be randomly generated from a SplitRand. Implementations are expected to exploit splittability where possible.

SplitRng
A trait for splittable pseudo random generators.

Type Definitions

Prf
The pseudo-random functions of a generic Split RNG.

Split
A wrapper that generically adds splittability to RNGs.