Crate rand_split

Splittable pseudo-random number generators.

These are generators that support a split operations that forks one or more "child" generators off the parent. Children's initial states depend on the source RNG's state, but they produce outputs that are uncorrelated to each other. This can be useful in some contexts:

Functional programs sometimes make good use of splittable RNGs. The Haskell QuickCheck library is a notable example that uses a splittable, pure-functional PRNG to support random generation of deterministic functions.

Parallel programs may benefit from a splittable PRNG. The advantage isn't necessarily performance but rather reproducibility; it makes it easier to guarantee that using the same seed yields the same results even if the execution order is nondeterministic.

Motivation

Splittable PRNGs offer superior reproducibility in some scenarios. Consider a case where you are randomly generating a pair of Rand instances with a conventional, i.e. sequential PRNG:

extern crate rand;
use rand::{Rng, SeedableRng, XorShiftRng, Rand, thread_rng};

// Make two PRNGs from the same seed.
let seed: [u32; 4] = thread_rng().gen();
let mut rng0: XorShiftRng = SeedableRng::from_seed(seed);
let mut rng1: XorShiftRng = SeedableRng::from_seed(seed);
 
type T0 = [u64; 16];
type T1 = [u64; 32];
let (_, a): (T0, T0) = Rand::rand(&mut rng0);
let (_, b): (T1, T0) = Rand::rand(&mut rng1);

// This should fail, unless you're crazy unlucky:
assert_eq!(a, b); 

With a sequential generator, the way this is generated typically is so:

The rng starts at some initial state. Call this state s0.

We use rng to generate a random a. This leaves the rng in some state, call it s1.

We reuse rng to generate a random b. This leaves the rng in some state, call it s2.

Now, if the logic that generates the a value changes, then states s1 may end up different even if we start from the same s0, which means that the value generated for b may also change.

With a splittable PRNG, you can generate the tuple's elements with independent PRNGs split off the parent, so that changes to the logic that generates one element cannot affect the other. This can be illustrated with our flagship unit test:

extern crate rand;
extern crate rand_split;

use rand::{Rng, Isaac64Rng, thread_rng};
use rand_split::{SplitRng, SplitPrf, SplitRand, Split, Prf};

// We will be using the `rand` crate's `Isaac64Rng`, but
// wrapped with a `Split` wrapper that adds splittability
// on top of it.
type OurRng = Split<Isaac64Rng>;
type OurPrf = Prf<Isaac64Rng>;

// The library's RNGs have `Rand` instances, so we can get
// a randomly seeded RNG this way:
let mut rng: OurRng = thread_rng().gen();

// We split off a **pseudo-random function** ("PRF") from
// the RNG.  PRFs implement the `SplitPrf` trait.
let prf: OurPrf = rng.splitn();
 
// PRFs serve as factories that construct further `SplitRng`s.
// So now we pick a random index and call the PRF four times
// with that index to get four new RNGs.
let i: u32 = rng.gen();
let mut ra: OurRng = prf.call(i);
let mut rb: OurRng = prf.call(i);
let mut rc: OurRng = prf.call(i);
let mut rd: OurRng = prf.call(i);
 
// A PRF is a function that captures a "frozen" state from
// its parent RNG, and constructs further RNG instances 
// whose initial states depend only on that frozen state and 
// the index supplied to `call`.  So the four RNGs we just
// constructed are guaranteed to be in the same initial state.
// And behold, they produce identical outputs!
assert!(iter_eq(ra.gen_ascii_chars().take(100),
                rb.gen_ascii_chars().take(100)));
assert!(iter_eq(rc.gen_ascii_chars().take(100),
                rd.gen_ascii_chars().take(100)));
assert!(iter_eq(ra.gen_ascii_chars().take(100),
                rc.gen_ascii_chars().take(100)));
assert!(iter_eq(rb.gen_ascii_chars().take(100),
                rd.gen_ascii_chars().take(100)));
 
// Now for the main course.  We pick two distinct types that
// implement the `SplitRand` trait.  We choose them so that
// generating a value of either type advances the state of a
// sequential RNG by a different amount than the other.
type T0 = [u64; 16];
type T1 = [u64; 32];
 
for _ in 0..100 {
    // Then we use our four initially-identical RNGs to
    // generate tuples representing all four combinations of
    // our two element types:
    let (a0, a1): (T0, T0) = SplitRand::split_rand(&mut ra);
    let (b0, b1): (T0, T1) = SplitRand::split_rand(&mut rb);
    let (c0, c1): (T1, T0) = SplitRand::split_rand(&mut rc);
    let (d0, d1): (T1, T1) = SplitRand::split_rand(&mut rd);
     
    // And here we test that the value of each element
    // generated depends on its type and its position within
    // its pair, but not on what was generated for the other
    // element.
    assert_eq!(a0, b0);
    assert_eq!(a1, c1);
    assert_eq!(b1, d1);
    assert_eq!(c0, d0);
    assert!(a0 != a1);
    assert!(a0 != c1);
    assert!(d0 != d1);
 
    // Finally, note that we're doing this inside of a loop
    // and reusing the same four RNGs on each iteration.  So 
    // at this point all four generators must end the same
    // state or subsequent iterations will fail.
}

This can be useful in programs that generate complex pseudo-random data from fixed seeds, because it means that changes in one location of the program can be made much less likely to affect the results produced in others. It can also be useful in parallel programs, because it can be used to provide deterministic pseudo-random generation in spite of non-deterministic execution order.

References

Claessen, Koen and Michał H. Pałka. 2013. "Splittable Pseudorandom Number Generators using Cryptographic Hashing." Haskell '13: Proceedings of the 2013 ACM SIGPLAN symposium on Haskell, pp. 47-58.

Schaathun, Hans Georg. 2015. "Evaluation of Splittable Pseudo-Random Generators." Journal of Functional Programming, Vol. 25.

The Haskell tf-random library.

Modules

generic
A construction that turns a pair of a splittable and a sequential PRNG into a splittable PRNG. The intent of this is to gain the splittability of the former but retain the sequential generation speed of the latter.

siprng
A splittable pseudo-random number generator based on the SipHash function. This is not intended to be a cryptographically secure PRNG.

Macros

split_rand_seq_impl!
A macro that implements SplitRand sequentially for any type that has a Rand implementation, simply by using that. This is meant to be used for "atomic" types whose generation doesn't benefit from splittability.

Structs

Seq
A newtype wrapper to add a SplitRand implementation to Rand types. This just does the same thing as the base type's Rand one does.

Traits

SplitPrf
Pseudo-random functions ("PRFs") generated off a SplitRng.

SplitRand
A type that can be randomly generated from a SplitRand. Implementations are expected to exploit splittability where possible.

SplitRng
A trait for splittable pseudo random generators.

Type Definitions

Prf
The pseudo-random functions of a generic Split RNG.

Split
A wrapper that generically adds splittability to RNGs.